Refactor: Rename rl to envs for consistency with highway-env naming conventions
This update aligns the naming scheme of the repository with that of highway-env, improving clarity and standardization across related projects.
