+++ replicantdrivesim/envs/__init__.py
@@ -1,4 +1,4 @@
-# rl/__init__.py
+# envs/__init__.py
 from .environment import CustomUnityMultiAgentEnv
 from .unity_env_resource import CustomSideChannel
 from .unity_env_resource import create_unity_env
